+++ cli/azd/pkg/infra/provisioning/bicep/prompt_test.go
@@ -279,6 +279,7 @@ func TestPromptForParametersLocation(t *testing.T) {
 		env,
 		mockContext.Console,
 		accountManager,
+		nil,
 		p.resourceService,
 		cloud.AzurePublic(),
 	)